Refactoring the S3SnaposhotStore class and SnapshotKeySupport trait to allow custom implementations of the SnapshotKeySupport functionality.
S3SnaposhotStore now takes a new config setting of 'key-generator-class'
that it uses to dynamically load an instance of the specified class.

A new class named DefaultSnapshotKeySupport replicates the original
functionality and is loaded by default if no 'key-generator-class' is
specified (so old code will continue to work as before).

A new PipePathKeySupport class has been created as a demonstration of
how to create a custom implementation.